BUGS trying to get item combination working. alpha channels are a little more complicated than I had anticipated.
The problem seems to be that the blitting function either uses the source alpha channel or the destination alpha channel. there isn't any intelligent selection.
I think I'm going to need to write my own blitting function (that will be much slower the the original blitting program) that handles combining different images

The problem with the current blitting program is that if it finds a source pixel with alpha of 0, it blitts the destination as that value. So, in effect, it just overwrites the destination as a completely transparent pixel. This should not be the case. it should simply skip that particular source pixel as it contributes nothing to the image.
(if there is a pixel with an alpha of 0, then it will simply skip blitting it.)