trying to seperate sentences into words. having issues. Maybe I need to try a different approach. Maybe I need to chop up the string into seperate strings and store them in different character arrays. That might be the easiest to work with. It sure would be a lot less screwing around with all of these numbers that keep track of where strings begin and end. It would make it nice and clean. I might want to make a function to do this. And maybe it would do this in general. maybe it could take a string and convert it into an array of strings (with each string being a maximum character width and there being some maximum number of strings that the function will break up the original string into. these would be input parameters.)
Yes. I'll have to do something along these lines.
Ok. I just wrote up a function outline with input variables and a lot of documentation. Now that the exterior of the function is good to go, all that is left to do now is to "connect the internal wires" of the function to make it preform operations on the data it is given and return a transformed paragraph.