modifying the debugging/error system.
as a convention, for debugging and error logs, every entry will start with two newlines to seperate it from the previous entry. This is done so that the errors/debugging output from one function is easily discernible from that of another.

This isn't overly important, but it makes the debug/error logs look nicer and more organized.