added game_time() function. working on player physics again.
game_time() returns the current time in milliseconds in the game you are. this ignores time that passes when the game is paused.

the player's position is currently discretized, but I added two members to the playerData structure:
    - x_pos_f
    - y_pos_f
these members keep track of the floating point location of the player. They aren't absolute. They are relative. So the player's "true" position is
    player.x_pos + player.x_pos_f;